requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pypi_0
anyio=4.5.2=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
array-record=0.4.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.13.3=pypi_0
bleach=6.1.0=pyhd8ed1ab_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
clang=5.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.2.2=hbe64b41_10
cudnn=8.9.2.26=cuda11_0
cycler=0.12.1=pypi_0
cython=3.1.1=pypi_0
debugpy=1.8.13=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pyhd8ed1ab_0
dm-tree=0.1.8=pypi_0
einops=0.7.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
et-xmlfile=2.0.0=pypi_0
etils=1.3.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.49.0=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2024.3.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.28.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.1.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.21.4=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.12.3=pypi_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.1.2=pyhd8ed1ab_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.3.2=pypi_0
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jupyter=1.0.0=pyhd8ed1ab_10
jupyter-client=8.6.1=pypi_0
jupyter-contrib-core=0.4.2=pypi_0
jupyter-contrib-nbextensions=0.7.0=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-nbextensions-configurator=0.6.3=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=5.7.1=py38h578d9bd_0
jupyterlab=4.3.6=pypi_0
jupyterlab-language-pack-zh-cn=4.1.post0=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab_widgets=3.0.10=pyhd8ed1ab_0
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=5.1.0=pypi_0
markdown=3.5.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.3=pypi_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
neurocombat-sklearn=0.1.3=pypi_0
notebook=7.1.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.19.5=pypi_0
nvidia-nccl-cu12=2.26.5=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=1.2.0=pypi_0
pandoc=2.19.2=ha770c72_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
patsy=1.0.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.3.6=pypi_0
plotly=5.24.1=pypi_0
pmdarima=2.0.4=pypi_0
prometheus-client=0.21.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.50=pypi_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=3.20.3=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure-eval=0.2.3=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.22=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.1.2=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=3.3.0=pypi_0
python_abi=3.8=2_cp38
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=pypi_0
qtconsole-base=5.5.1=pyha770c72_0
qtpy=2.4.1=pyhd8ed1ab_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.20.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.11.0=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.3.2=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
spektral=1.3.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.1=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-addons=0.14.0=pypi_0
tensorflow-datasets=4.9.2=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-gpu=2.6.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
tensorflow-metadata=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.18.1=pypi_0
testpath=0.6.0=pyhd8ed1ab_0
tf-keras=2.15.0=pypi_0
threadpoolctl=3.3.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.38.2=pypi_0
typeguard=2.13.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.13.0=pypi_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=pypi_0
validators=0.23.2=pypi_0
vit-keras=0.1.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.8.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
widgetsnbextension=4.0.10=pyhd8ed1ab_0
wrapt=1.12.1=pypi_0
xgboost=2.1.4=pypi_0
xz=5.4.6=h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_0
